+++ src/routes/memberList.js
@@ -7,19 +7,19 @@ import { faArrowsRotate, faUserGroup } from '@fortawesome/free-solid-svg-icons';
 
 import TimeAgo from '../components/timeago';
 import CustomTable from "../components/table";
-import { makeRequestsAuto, customAxios as axios, getAuthToken } from '../functions';
+import { makeRequestsAuto, customAxios as axios, getAuthToken, removeNUEValues } from '../functions';
 import UserCard from '../components/usercard';
 
 var vars = require("../variables");
 
 const columns = [
-    { id: 'userid', label: 'User ID' },
-    { id: 'user', label: 'User' },
-    { id: 'discordid', label: 'Discord ID' },
+    { id: 'userid', label: 'User ID', orderKey: 'userid', defaultOrder: 'asc' },
+    { id: 'user', label: 'User', orderKey: 'name', defaultOrder: 'asc' },
+    { id: 'discordid', label: 'Discord ID', orderKey: 'discordid', defaultOrder: 'asc' },
     { id: 'steamid', label: 'Steam ID' },
     { id: 'truckersmpid', label: 'TruckersMP ID' },
-    { id: 'joined', label: 'Joined' },
-    { id: 'last_seen', label: 'Last Seen' }
+    { id: 'joined', label: 'Joined', orderKey: 'join_timestamp', defaultOrder: 'asc' },
+    { id: 'last_seen', label: 'Last Seen', orderKey: 'last_seen', defaultOrder: 'desc' }
 ];
 
 const MemberList = () => {
@@ -30,6 +30,7 @@ const MemberList = () => {
     const [page, setPage] = useState(-1);
     const [pageSize, setPageSize] = useState(10);
     const [search, setSearch] = useState("");
+    const [listParam, setListParam] = useState({ order_by: "userid", order: "asc" });
 
     const [dialogOpen, setDialogOpen] = useState("");
     const [dialogButtonDisabled, setDialogButtonDisabled] = useState(false);
@@ -85,15 +86,17 @@ const MemberList = () => {
 
         let myPage = page;
         myPage === -1 ? myPage = 1 : myPage += 1;
+        
+        let processedParam = removeNUEValues(listParam);
 
         let [_userList] = [{}];
         if (search === "")
             [_userList] = await makeRequestsAuto([
-                { url: `${vars.dhpath}/member/list?order=desc&order_by=userid&page=${myPage}&page_size=${pageSize}`, auth: true },
+                { url: `${vars.dhpath}/member/list?order=desc&order_by=userid&page=${myPage}&page_size=${pageSize}&${new URLSearchParams(processedParam).toString()}`, auth: true },
             ]);
         else if (isNaN(search) || !isNaN(search) && (search.length < 17 || search.length > 19)) // not discord id
             [_userList] = await makeRequestsAuto([
-                { url: `${vars.dhpath}/member/list?query=${search}&order=desc&order_by=userid&page=${myPage}&page_size=${pageSize}`, auth: true },
+                { url: `${vars.dhpath}/member/list?query=${search}&order=desc&order_by=userid&page=${myPage}&page_size=${pageSize}&${new URLSearchParams(processedParam).toString()}`, auth: true },
             ]);
         else if (!isNaN(search) && search.length >= 17 && search.length <= 19) { // is discord id
             let [_userProfile] = await makeRequestsAuto([
@@ -117,13 +120,13 @@ const MemberList = () => {
 
         const loadingEnd = new CustomEvent('loadingEnd', {});
         window.dispatchEvent(loadingEnd);
-    }, [theme, page, pageSize, search]);
+    }, [theme, page, pageSize, search, listParam]);
     useEffect(() => {
         doLoad();
     }, [doLoad]);
 
     return <>
-        <CustomTable name={<><FontAwesomeIcon icon={faUserGroup} />&nbsp;&nbsp;Members&nbsp;&nbsp;<IconButton onClick={() => { setDialogOpen("sync-profile"); }}><FontAwesomeIcon icon={faArrowsRotate} /></IconButton></>} titlePosition="top" columns={columns} data={userList} totalItems={totalItems} rowsPerPageOptions={[10, 25, 50, 100, 250]} defaultRowsPerPage={pageSize} onPageChange={setPage} onRowsPerPageChange={setPageSize} onSearch={(content) => { setPage(-1); setSearch(content); }} searchHint="Search by username or discord id" />
+        <CustomTable name={<><FontAwesomeIcon icon={faUserGroup} />&nbsp;&nbsp;Members&nbsp;&nbsp;<IconButton onClick={() => { setDialogOpen("sync-profile"); }}><FontAwesomeIcon icon={faArrowsRotate} /></IconButton></>} order={listParam.order} orderBy={listParam.order_by} onOrderingUpdate={(order_by, order) => { setListParam({ ...listParam, order_by: order_by, order: order }); }} titlePosition="top" columns={columns} data={userList} totalItems={totalItems} rowsPerPageOptions={[10, 25, 50, 100, 250]} defaultRowsPerPage={pageSize} onPageChange={setPage} onRowsPerPageChange={setPageSize} onSearch={(content) => { setPage(-1); setSearch(content); }} searchHint="Search by username or discord id" />
         <Dialog open={dialogOpen === "sync-profile"} onClose={() => { if (!dialogButtonDisabled) setDialogOpen(""); }}>
             <DialogTitle><FontAwesomeIcon icon={faArrowsRotate} />&nbsp;&nbsp;Sync Profiles</DialogTitle>
             <DialogContent>